requirements.txt
langchain==0.1.0
langchain-openai==0.0.5
langchain-community==0.0.10
openai==1.3.0
python-dotenv==1.0.0
pydantic==2.5.0
typing-extensions==4.8.0
requests==2.31.0
colorama==0.4.6
rich==13.7.0 